refactor : fix all types add them to the types folder
