+++ src/types/userTypes.ts
@@ -7,3 +7,12 @@ export interface UserState {
 	setToken: (token: string) => void; // Add setToken function
   }
   
+export  interface UserProfile {
+    id: string;
+    username: string;
+    imagePath: string;
+    bio: string;
+    followersCount: number;
+    followingCount: number;
+    doneChallengesCount: number;
+  }
